requirements.txt
streamlit
langchain
openai
faiss-cpu
PyPDF2
tiktoken
python-dotenv